+++ src/ifpa_api/resources/__init__.py
@@ -8,7 +8,7 @@ from ifpa_api.resources.director import DirectorClient
 from ifpa_api.resources.player import PlayerClient
 from ifpa_api.resources.rankings import RankingsClient
 from ifpa_api.resources.reference import ReferenceClient
-from ifpa_api.resources.series import SeriesClient, SeriesHandle
+from ifpa_api.resources.series import SeriesClient
 from ifpa_api.resources.tournament import TournamentClient
 
 __all__ = [
@@ -17,6 +17,5 @@ __all__ = [
     "RankingsClient",
     "ReferenceClient",
     "TournamentClient",
-    "SeriesHandle",
     "SeriesClient",
 ]